+++ components/home/ItemManage.vue
@@ -1,0 +1,13 @@
+<template>
+    <div></div>
+</template>
+
+<style scoped>
+
+</style>
+
+<script>
+export default {
+    
+}
+</script>
